+++ public/static/home/js/global.js
@@ -270,7 +270,7 @@ go.calcTotalAmount = function (num, price) {
         return res;
     }
 
-    num = parseInt(num);
+    num = parseFloat(num);
     if (isNaN(num)) {
         return res;
     }
